requirements.txt
fastapi
uvicorn
requests
redis
python-dotenv
pillow
pydantic
wechatpy
cryptography

cryptography
psycopg2-binary
httpx
